requirements-ci.txt
pytest
pillow
# Don't need pre-commit in CI
# pre-commit
# Install directly to venv in CI
black==22.3.0
flake8==4.0.1
isort==5.10.1
